Refactor ComplexDoubleArray
