+++ Sources/VecLab/Types/ComplexArray/ComplexDoubleArray+Operators.swift
@@ -12,6 +12,7 @@ extension ComplexDoubleArray {
     public static func += (lhs: inout ComplexDoubleArray, rhs: ComplexDoubleArray) {
         lhs = lhs + rhs
     }
+    
 /*
     public static func += (lhs: inout ComplexDoubleArray, rhs: [Double]) {
         lhs = lhs + rhs
